# Fresh-Dating/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Trust Dating API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile/me:
    get:
      operationId: getMyProfile
      tags: [profile]
      summary: Get my profile
      responses:
        "200":
          description: My profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateMyProfile
      tags: [profile]
      summary: Update my profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /profile/me/now-items:
    post:
      operationId: createNowItem
      tags: [profile]
      summary: Add a Now tab item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNowItemRequest"
      responses:
        "201":
          description: Now item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NowItem"
  /profile/me/now-items/{nowItemId}:
    delete:
      operationId: deleteNowItem
      tags: [profile]
      summary: Delete a Now item
      parameters:
        - name: nowItemId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /profiles/{userId}:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get a user's profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /users/{userId}/block:
    post:
      operationId: blockUser
      tags: [users]
      summary: Block a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Blocked
  /users/{userId}/report:
    post:
      operationId: reportUser
      tags: [users]
      summary: Report a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportRequest"
      responses:
        "204":
          description: Reported
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        userId:
          type: string
        displayName:
          type: string
        bio:
          type: string
        age:
          type: integer
        genderIdentity:
          type: string
        orientation:
          type: string
        relationshipIntent:
          type: string
        verificationStatus:
          type: string
          enum: [unverified, pending, verified]
        freshnessScore:
          type: number
        freshnessLabel:
          type: string
          enum: [active_today, updated_this_week, recent_media_this_month]
        profileCompletenessScore:
          type: number
        photos:
          type: array
          items:
            $ref: "#/components/schemas/ProfilePhoto"
        prompts:
          type: array
          items:
            $ref: "#/components/schemas/PromptAnswer"
        nowItems:
          type: array
          items:
            $ref: "#/components/schemas/NowItem"
        premiumTier:
          type: string
          enum: [free, premium]
        lastActiveAt:
          type: string
          format: date-time
        locationCity:
          type: string
        distanceKm:
          type: number
      required:
        - userId
        - displayName
        - verificationStatus
        - freshnessScore
        - profileCompletenessScore
        - photos
        - prompts
        - nowItems
        - premiumTier
    ProfilePhoto:
      type: object
      properties:
        photoId:
          type: string
        url:
          type: string
        isPrimary:
          type: boolean
        freshnessLabel:
          type: string
        uploadedAt:
          type: string
          format: date-time
      required:
        - photoId
        - url
        - isPrimary
    PromptAnswer:
      type: object
      properties:
        promptAnswerId:
          type: string
        promptText:
          type: string
        answerType:
          type: string
          enum: [text, audio, video]
        contentText:
          type: string
        updatedAt:
          type: string
          format: date-time
      required:
        - promptAnswerId
        - promptText
        - answerType
    NowItem:
      type: object
      properties:
        nowItemId:
          type: string
        mediaType:
          type: string
          enum: [photo, audio, video, text]
        caption:
          type: string
        url:
          type: string
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
      required:
        - nowItemId
        - mediaType
        - createdAt
    UpdateProfileRequest:
      type: object
      properties:
        displayName:
          type: string
        bio:
          type: string
        relationshipIntent:
          type: string
        locationCity:
          type: string
    CreateNowItemRequest:
      type: object
      properties:
        mediaType:
          type: string
          enum: [photo, audio, video, text]
        caption:
          type: string
        url:
          type: string
      required:
        - mediaType
...
    ReportRequest:
      type: object
      properties:
        category:
          type: string
          enum: [fake_profile, harassment, inappropriate_content, spam_scam, underage_concern, other]
        details:
          type: string
      required:
        - category
...